Cierra contrato de spans y diagnósticos del lexer

+++ include/token.h
@@ -112,10 +112,10 @@ typedef struct Token {
     MilenaTokenType type;
     char lexeme[MAX_TOKEN_LEN];
     /* One-based start and exclusive end coordinates; offsets are source bytes. */
-    int line;
-    int column;
-    int end_line;
-    int end_column;
+    size_t line;
+    size_t column;
+    size_t end_line;
+    size_t end_column;
     size_t start_offset;
     size_t end_offset;
     double number_value;
